Add automatic-corner detection for test correction
There're test that were bad scaned. As a result, those tests were
not analized since the algorithm look for static positions in terms
of (x,y) coordinates.

Enhancement - It was implemented dynamic solution that look
for four black squares in each test in oder to draw the template
correction depending on that four black squares positions  and
detect the correct answers of those test bad scaned by run the
algorithm. Additionally, in case is not recognized one of the four
squares the algorithm can calculate the fourth missing position by
appling math calculations. In other cases, like if is missing more
than 2 black squares, those test with not be corrected.